pyproject.toml
[tool.poetry]
name = "eventstore_grpc"
version = "0.1.0"
description = "EventStoreDB gRPC client."
authors = ["Stefano Frassetto <frassetto.stefano@gmail.com>"]
maintainers = ["Stefano Frassetto <frassetto.stefano@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/stefanondisponibile/eventstore_grpc"
repository = "https://github.com/stefanondisponibile/eventstore_grpc"
documentation = "https://github.com/stefanondisponibile/eventstore_grpc"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Topic :: Database"
]
packages = [
    {include = "eventstore_grpc", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
protobuf = "3.20.1"
grpcio = "*"
grpcio-status = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
black = "^22.3.0"
mypy = "^0.961"
flake8 = "^4.0.1"
grpcio-tools = "*"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
